plot_beam -> beam_plot.py
